services/nlp/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
boto3==1.34.0
groq==0.9.0
httpx==0.27.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
